get app running again
