+++ package.json
@@ -8,6 +8,7 @@
   ],
   "scripts": {
     "build": "pnpm run --filter=next-yak build",
+    "watch": "pnpm run --filter=next-yak watch",
     "example": "pnpm --filter=next-yak-example run dev",
     "test": "pnpm --filter=next-yak run \"/test($|:types)/\"",
     "test:watch": "pnpm --filter=next-yak run test:watch",
