+++ packages/example/app/layout.tsx
@@ -1,22 +1,16 @@
-import "./globals.css";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "CSS in JS POC",
-  description: "Styled Components API with ReactServerComponents",
-};
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode;
+  children: React.ReactNode
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body>{children}</body>
     </html>
-  );
+  )
 }
