--- packages/example/app/page.module.css
@@ -1,8 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding: 6rem;
-  gap: 2rem;
-}
